Add some utils for taggeth command and restructure config
The taggeth command, once introduced, allows Aschenputtel
to store deleted messages that contained mentions of
users or roles as a log.

The command itself is not implemented yet, but the logging
already works.

This reuired the config to be restructured, as now every
command has its own permission subnode, instead of
one big permission node.